Update sidebar
The sidebar link section has been removed, and the link bar (FAQs,
Resources, Community) has been moved to its place. This should make the
sidebar less redundant for oldreddit users, as links will only appear in
one place. And it should be better for mobile users and those who turn
off CSS, as the important sidebar information is now at the top.

The sidebar has also been updated with the newest headerbar links, but
those will soon be out of date.

There are also a handful of small changes, such as grammar fixes, a link
to the DT where it's mentioned, and changing the flair button URL so RES
doesn't show tooltip on hover.