Rename files to fix flair collisions
When multiple flairs have the same name, SCSS doesn't know what to do,
and one of them breaks. This fixes issues with Borlaug and Marshall.
Additionally, I moved the _unused and _politician flairs to a separate
directory so I can glob the flairs directory for all in-use flairs,
which will be useful for the emoji migration.