Reformat sidebar
The positions and styling of the buttons (get a flair... view
effortposts), the link bar (faqs... community) and featured posts have
been overhauled. This was done to allow for improvements to the layout
and style of the sidebar for those not using CSS.

Additionally, it has moved the featured post text from the CSS to the
sidebar. This makes it easier to edit for the other mods and there won't
be any more "update featured post" commits.